Refactoring code to reflect rule storing
