+++ src/main/java/com/takatsuka/web/math/Expression/AST.java
@@ -1,4 +1,4 @@
-package com.takatsuka.web.Expression;
+package com.takatsuka.web.math.Expression;
 
 import java.util.Queue;
