+++ src/main/java/com/takatsuka/web/math/rules/RuleLoader.java
@@ -1,0 +1,4 @@
+package com.takatsuka.web.math.rules;
+
+public class RuleLoader {
+}
